backend/requirements.txt
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.2
requests>=2.31.0
scikit-learn>=1.3.0
gitpython>=3.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
pymongo>=4.5.0
dnspython>=2.4.0
gunicorn>=21.2.0

# Optional AI integration
google-generativeai>=0.3.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-flask>=1.2.0
